Extract HTTP-related settings into HttpConfiguration
