+++ graylog2-server/src/main/java/org/graylog2/periodical/NodePingThread.java
@@ -16,10 +16,10 @@
  */
 package org.graylog2.periodical;
 
-import org.graylog2.Configuration;
 import org.graylog2.cluster.Node;
 import org.graylog2.cluster.NodeNotFoundException;
 import org.graylog2.cluster.NodeService;
+import org.graylog2.configuration.HttpConfiguration;
 import org.graylog2.notifications.Notification;
 import org.graylog2.notifications.NotificationImpl;
 import org.graylog2.notifications.NotificationService;
@@ -33,28 +33,25 @@ import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 
-/**
- * @author Lennart Koopmann <lennart@torch.sh>
- */
 public class NodePingThread extends Periodical {
 
     private static final Logger LOG = LoggerFactory.getLogger(NodePingThread.class);
     private final NodeService nodeService;
     private final NotificationService notificationService;
     private final ActivityWriter activityWriter;
-    private final Configuration configuration;
+    private final HttpConfiguration httpConfiguration;
     private final ServerStatus serverStatus;
 
     @Inject
     public NodePingThread(NodeService nodeService,
                           NotificationService notificationService,
                           ActivityWriter activityWriter,
-                          Configuration configuration,
+                          HttpConfiguration httpConfiguration,
                           ServerStatus serverStatus) {
         this.nodeService = nodeService;
         this.notificationService = notificationService;
         this.activityWriter = activityWriter;
-        this.configuration = configuration;
+        this.httpConfiguration = httpConfiguration;
         this.serverStatus = serverStatus;
     }
 
@@ -63,12 +60,12 @@ public class NodePingThread extends Periodical {
         final boolean isMaster = serverStatus.hasCapability(ServerStatus.Capability.MASTER);
         try {
             Node node = nodeService.byNodeId(serverStatus.getNodeId());
-            nodeService.markAsAlive(node, isMaster, configuration.getRestTransportUri());
+            nodeService.markAsAlive(node, isMaster, httpConfiguration.getRestTransportUri());
         } catch (NodeNotFoundException e) {
             LOG.warn("Did not find meta info of this node. Re-registering.");
             nodeService.registerServer(serverStatus.getNodeId().toString(),
                     isMaster,
-                    configuration.getRestTransportUri(),
+                    httpConfiguration.getRestTransportUri(),
                     Tools.getLocalCanonicalHostname());
         }
         try {
