+++ graylog2-server/src/main/java/org/graylog2/rest/resources/HelloWorldResource.java
@@ -19,7 +19,7 @@ package org.graylog2.rest.resources;
 import com.codahale.metrics.annotation.Timed;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
-import org.graylog2.Configuration;
+import org.graylog2.configuration.HttpConfiguration;
 import org.graylog2.plugin.Version;
 import org.graylog2.plugin.cluster.ClusterConfigService;
 import org.graylog2.plugin.cluster.ClusterId;
@@ -42,15 +42,15 @@ import static java.util.Objects.requireNonNull;
 public class HelloWorldResource extends RestResource {
     private final NodeId nodeId;
     private final ClusterConfigService clusterConfigService;
-    private final Configuration configuration;
+    private final HttpConfiguration httpConfiguration;
 
     @Inject
     public HelloWorldResource(NodeId nodeId,
                               ClusterConfigService clusterConfigService,
-                              Configuration configuration) {
+                              HttpConfiguration httpConfiguration) {
         this.nodeId = requireNonNull(nodeId);
         this.clusterConfigService = requireNonNull(clusterConfigService);
-        this.configuration = configuration;
+        this.httpConfiguration = requireNonNull(httpConfiguration, "configuration");
     }
 
     @GET
@@ -72,8 +72,8 @@ public class HelloWorldResource extends RestResource {
     @ApiOperation(value = "Redirecting to web console if it runs on same port.")
     @Produces({MediaType.TEXT_HTML, MediaType.APPLICATION_XHTML_XML})
     public Response redirectToWebConsole() {
-        if (configuration.isRestAndWebOnSamePort()) {
-            final URI target = URI.create(configuration.getWebPrefix());
+        if (httpConfiguration.isRestAndWebOnSamePort()) {
+            final URI target = URI.create(httpConfiguration.getWebPrefix());
             return Response
                 .temporaryRedirect(target)
                 .build();
