+++ graylog2-server/src/main/java/org/graylog2/shared/rest/resources/RestResource.java
@@ -22,6 +22,7 @@ import com.fasterxml.jackson.jaxrs.cfg.EndpointConfigBase;
 import com.fasterxml.jackson.jaxrs.cfg.ObjectWriterInjector;
 import com.fasterxml.jackson.jaxrs.cfg.ObjectWriterModifier;
 import org.apache.shiro.subject.Subject;
+import org.graylog2.configuration.HttpConfiguration;
 import org.graylog2.indexer.IndexSet;
 import org.graylog2.indexer.IndexSetRegistry;
 import org.graylog2.plugin.BaseConfiguration;
@@ -54,7 +55,7 @@ public abstract class RestResource {
     protected UserService userService;
 
     @Inject
-    private BaseConfiguration configuration;
+    private HttpConfiguration configuration;
 
     @Context
     SecurityContext securityContext;
